# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "互动直播API",
        "title": "互动直播API",
        "termsOfService": "https://www.qiniu.com",
        "contact": {
            "name": "qlive developer",
            "url": "https://github.com/qrtc/qlive"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/enter_room": {
            "post": {
                "description": "Audience enters room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "Audience enters room",
                "operationId": "enter-room",
                "parameters": [
                    {
                        "description": "Audience enters room",
                        "name": "EnterRoomRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.EnterRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.EnterRoomResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/im_user_token": {
            "post": {
                "description": "User Gets user-token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "Get user token",
                "operationId": "get-user-token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.IMTokenResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/leave_room": {
            "post": {
                "description": "Audience leaves room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "Audience leaves room",
                "operationId": "leave-room",
                "parameters": [
                    {
                        "description": "Audience leaves room",
                        "name": "LeaveRoomArgs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.LeaveRoomArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "User log in with sms code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "User log in",
                "operationId": "log-in",
                "parameters": [
                    {
                        "type": "string",
                        "description": "type of user logs in",
                        "name": "logintype",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "user's phone number and sms code",
                        "name": "SMSLoginArgs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.SMSLoginArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "Online user logs out",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "User logs out",
                "operationId": "log-out",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/profile": {
            "post": {
                "description": "User updates personal profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "User updates profile",
                "operationId": "update-profile",
                "parameters": [
                    {
                        "description": "User updates personal profile",
                        "name": "updatedProfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.UpdateProfileArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/refresh_room": {
            "post": {
                "description": "Anchor refreshes room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "Anchor refreshes room",
                "operationId": "refresh-room",
                "parameters": [
                    {
                        "description": "Anchor requests to refresh room",
                        "name": "RefreshRoomArgs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.RefreshRoomArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.RefreshRoomResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/rooms": {
            "get": {
                "description": "if can_pk set in url,server will returns rooms that can pk.Otherwise,all rooms will be returned",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "List rooms",
                "operationId": "list-rooms",
                "parameters": [
                    {
                        "type": "string",
                        "description": "only list rooms that can pk",
                        "name": "can_pk",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.ListRoomsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Anchor closes room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "Anchor closes room",
                "operationId": "close-room",
                "parameters": [
                    {
                        "description": "Anchor requests to close room",
                        "name": "CloseRoomArgs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.CloseRoomArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/rooms/{roomID}": {
            "get": {
                "description": "Get room by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "Get room by id",
                "operationId": "get-room-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": " Get room by id",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.GetRoomResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Anchor updates room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "Anchor updates room",
                "operationId": "update-room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room id",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Anchor requests to update room",
                        "name": "UpdateRoomArgs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.UpdateRoomArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/send_sms_code": {
            "post": {
                "description": "Send SMS code to user and the code will survive for ten minutes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qlive api"
                ],
                "summary": "Send SMS code to user",
                "operationId": "send-sms-code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Send sms code to user's phone number",
                        "name": "phone_number",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "自定义错误码。",
                    "type": "integer"
                },
                "message": {
                    "description": "Message 错误消息。",
                    "type": "string"
                },
                "requestID": {
                    "description": "请求ID。",
                    "type": "string"
                },
                "summary": {
                    "description": "Summary",
                    "type": "string"
                }
            }
        },
        "protocol.CloseRoomArgs": {
            "type": "object",
            "properties": {
                "roomID": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "protocol.CreateRoomArgs": {
            "type": "object",
            "properties": {
                "roomName": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "protocol.EnterRoomRequest": {
            "type": "object",
            "properties": {
                "roomID": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "protocol.EnterRoomResponse": {
            "type": "object",
            "properties": {
                "creator": {
                    "description": "直播间创建者信息。",
                    "type": "object",
                    "$ref": "#/definitions/protocol.UserInfo"
                },
                "imGroup": {
                    "description": "IMGroup IM聊天室信息。",
                    "type": "object",
                    "$ref": "#/definitions/protocol.IMChatRoom"
                },
                "imUser": {
                    "description": "IMUser IM聊天用户信息。",
                    "type": "object",
                    "$ref": "#/definitions/protocol.IMUser"
                },
                "pkAnchor": {
                    "description": "PKAnchorID 若正在PK，返回PK主播的信息。未在PK时该字段为空。",
                    "type": "object",
                    "$ref": "#/definitions/protocol.UserInfo"
                },
                "playURL": {
                    "description": "直播间的拉流观看地址。",
                    "type": "string"
                },
                "roomID": {
                    "type": "string"
                },
                "roomName": {
                    "type": "string"
                },
                "status": {
                    "description": "TODO：添加创建者的昵称/性别等信息？\nStatus 房间的状态，单人直播中/PK中",
                    "type": "string"
                }
            }
        },
        "protocol.GetRoomResponse": {
            "type": "object",
            "properties": {
                "audienceNumber": {
                    "description": "只返回观众人数",
                    "type": "integer"
                },
                "creator": {
                    "description": "创建者的用户ID",
                    "type": "object",
                    "$ref": "#/definitions/protocol.UserInfo"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pkAnchor": {
                    "description": "PKAnchor 若房间PK中，返回PK主播的信息。",
                    "type": "object",
                    "$ref": "#/definitions/protocol.UserInfo"
                },
                "playURL": {
                    "description": "TODO：添加创建者的昵称/性别等信息？\nWatchURL 观看地址",
                    "type": "string"
                },
                "status": {
                    "description": "房间状态，单人直播中/PK中。",
                    "type": "string"
                }
            }
        },
        "protocol.IMChatRoom": {
            "type": "object"
        },
        "protocol.IMTokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "protocol.IMUser": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "protocol.LeaveRoomArgs": {
            "type": "object",
            "properties": {
                "roomID": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "protocol.ListRoomsResponse": {
            "type": "object",
            "properties": {
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/protocol.GetRoomResponse"
                    }
                }
            }
        },
        "protocol.LoginResponse": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "protocol.RefreshRoomArgs": {
            "type": "object",
            "properties": {
                "roomID": {
                    "type": "string"
                }
            }
        },
        "protocol.RefreshRoomResponse": {
            "type": "object",
            "properties": {
                "roomID": {
                    "type": "string"
                },
                "roomName": {
                    "type": "string"
                },
                "rtcRoom": {
                    "description": "RTCRoom 对应的RTC房间。",
                    "type": "string"
                },
                "rtcRoomToken": {
                    "description": "RTCRoomToken 创建/加入RTC房间的token。",
                    "type": "string"
                },
                "wsURL": {
                    "description": "WSURL websocket 服务器地址。",
                    "type": "string"
                }
            }
        },
        "protocol.SMSLoginArgs": {
            "type": "object",
            "properties": {
                "phoneNumber": {
                    "type": "string"
                },
                "smsCode": {
                    "type": "string"
                }
            }
        },
        "protocol.UpdateProfileArgs": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "protocol.UpdateRoomArgs": {
            "type": "object",
            "properties": {
                "roomName": {
                    "type": "string"
                }
            }
        },
        "protocol.UserInfo": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        }
    }
}